refactor initialization of store
